divided reducer
