+++ src/components/ingridient-details/ingridient-details.jsx
@@ -9,7 +9,6 @@ import { useSelector  } from 'react-redux'
 const IngridientDetails = () => {
   const currentItemToView = useSelector(state => state.app.currentItemToView)
   const order = useSelector(state => state.app.order)
-  // console.log(order)
   return (
     <>
       {
